requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
qdrant-client==1.7.3
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
